refactor: updated little glitches
