+++ package.json
@@ -103,7 +103,6 @@
     "sass-extract": "~2.1.0",
     "sass-extract-loader": "~1.1.0",
     "sass-loader": "~10.1.0",
-    "stencil-inline-svg": "^1.0.1",
     "style-loader": "~2.0.0",
     "typescript": "~4.1.2",
     "wait-on": "5.2.1",
