--- src/components/zen-icon/types.tsx
@@ -1,1 +1,0 @@
-export type IconSizes = 'md' | 'lg';
