+++ src/components/zen-icon/zen-icon.scss
@@ -5,7 +5,7 @@
   justify-content: center;
 }
 
-:host(.lg) {
+:host([size='lg']) {
   width: 1.5rem;
   height: 1.5rem;
 }
