+++ src/components/zen-icon/zen-icon.stories.mdx
@@ -22,8 +22,6 @@ export const StoryDefault = args => {
 
 Icon component wrapper
 
-## Component previews
-
 <br />
 
 ## Properties
