+++ src/components/zen-icon/zen-icon.tsx
@@ -1,7 +1,7 @@
 import { IconDefinition } from '@fortawesome/pro-light-svg-icons';
 import { Component, Host, h, Prop, Element } from '@stencil/core';
 import { renderIcon } from '../helpers/fa-icons';
-import { IconSizes } from './types';
+import { IconSizes } from '../helpers/types';
 
 @Component({
   tag: 'zen-icon',
@@ -12,12 +12,12 @@ export class ZenIcon {
   @Element() hostElement: HTMLZenIconElement;
 
   /** Size of the icon. */
-  @Prop() readonly size: IconSizes = 'md';
+  @Prop({ reflect: true }) readonly size: IconSizes = 'md';
 
-  /** Icon to be rendered. */
+  /** Icon data (js file) imported from Font Awesome SVG package. */
   @Prop() readonly icon: IconDefinition = null;
 
   render(): HTMLElement {
-    return <Host class={{ lg: this.size === 'lg' }}>{this.icon ? renderIcon(this.icon) : null}</Host>;
+    return <Host>{this.icon ? renderIcon(this.icon) : null}</Host>;
   }
 }
